+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # v11.4.2
 
+## Enhancements
+- [#213] Add ability to toggle Radiation and Poison immunity on character and NPC sheets by clicking the Radiation or Poison icon in the resistance sections on character sheets.
+
 ## Bugfixes
 - [#212] Missing material consumables: Asbestos, Bloatfly Gland, Blood Sac, Bloodleaf, Glowing Fungus, Hubflower, Radscorpion Stinger, Stingwing Barb